Changed the price into decimal Add Real Data's on Order Summary,removed the dummy paragraph @ home categories

+++ database/migrations/2017_05_24_134733_create_products_table.php
@@ -20,7 +20,7 @@ class CreateProductsTable extends Migration
             $table->integer('stock');
             $table->string('size');
             $table->string('color');
-            $table->bigInteger('price');
+            $table->decimal('price',11,2);
             $table->string('category');
             $table->string('brand_id');
             $table->timestamps();
